02 - MF0492_3 - Programación web en el entorno servido - 240h/UF1844 - Desarrollo de aplicaciones web en el entorno servidor - 90h/Java/PaquetesJava/src/servicio/PersonaService.java
package servicio;

import modelo.Persona;
import modelo.Empleado;


public class PersonaService {
    public void imprimirDetalles(Persona persona) {
        System.out.println("Nombre: " + persona.getNombre());
        System.out.println("Edad: " + persona.getEdad());
    }

    public void imprimirDetalles(Empleado empleado) {
        System.out.println("Nombre: " + empleado.getNombre());
        System.out.println("Edad: " + empleado.getEdad());
        System.out.println("Salario: " + empleado.getSalario());
        System.out.println("Puesto: " + empleado.getPuesto());
    }
}
